do on barchart
